remove grid, rename Infinite to InfiniteGrid
