+++ src/main.ts
@@ -1,5 +1,3 @@
 import App from "./app/app"
 
-(()=>{
-    new App()
-})()
+new App()
